pyproject.toml
[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "photo-organizer"
version = "0.2.0"
description = "Comprehensive toolkit for organizing photos, videos, and scans"
readme = "README.md"
authors = ["Ulises Chavarria"]
license = "MIT"
packages = [
    { include = "photo_organizer", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
Pillow = "^10.4.0"
pillow-heif = "^0.13.1"
exifread = "^3.3.1"
tqdm = "^4.66.0"
appdirs = "^1.4.4"

[tool.poetry.scripts]
photo-organizer = "photo_organizer.launcher:main"
